+++ scheduler/classes.c
@@ -1,7 +1,7 @@
 /*
  * Printer class routines for the CUPS scheduler.
  *
- * Copyright 2007-2014 by Apple Inc.
+ * Copyright 2007-2017 by Apple Inc.
  * Copyright 1997-2007 by Easy Software Products, all rights reserved.
  *
  * These coded instructions, statements, and computer programs are the
@@ -345,6 +345,13 @@ cupsdLoadAllClasses(void)
     {
       cupsdLogMessage(CUPSD_LOG_ERROR,
                       "Syntax error on line %d of classes.conf.", linenum);
+    }
+    else if (!_cups_strcasecmp(line, "PrinterId"))
+    {
+      if (value && (i = atoi(value)) > 0)
+        p->printer_id = i;
+      else
+        cupsdLogMessage(CUPSD_LOG_ERROR, "Bad PrinterId on line %d of classes.conf.", linenum);
     }
     else if (!_cups_strcasecmp(line, "UUID"))
     {
@@ -712,6 +719,9 @@ cupsdSaveAllClasses(void)
       cupsFilePrintf(fp, "<DefaultClass %s>\n", pclass->name);
     else
       cupsFilePrintf(fp, "<Class %s>\n", pclass->name);
+
+    if (pclass->printer_id)
+      cupsFilePrintf(fp, "PrinterId %d\n", pclass->printer_id);
 
     cupsFilePrintf(fp, "UUID %s\n", pclass->uuid);