+++ scheduler/printers.h
@@ -58,6 +58,7 @@ typedef struct cupsd_job_s cupsd_job_t;
 struct cupsd_printer_s
 {
   _cups_rwlock_t lock;			/* Concurrency lock for background updates */
+  int		printer_id;		/* Printer ID */
   char		*uri,			/* Printer URI */
 		*uuid,			/* Printer UUID */
 		*hostname,		/* Host printer resides on */
@@ -137,6 +138,8 @@ VAR ipp_t		*CommonData	VALUE(NULL);
 					/* Common printer object attrs */
 VAR cups_array_t	*CommonDefaults	VALUE(NULL);
 					/* Common -default option names */
+VAR int			NextPrinterId	VALUE(1);
+					/* Next printer-id value */
 VAR cups_array_t	*Printers	VALUE(NULL);
 					/* Printer list */
 VAR cupsd_printer_t	*DefaultPrinter	VALUE(NULL);